Fix issues with interface for direct/heap implementation
